+++ Address-Finder/Address-Finder/Address.swift
@@ -1,0 +1,17 @@
+//
+//  AddressStruct.swift
+//  Address-Finder
+//
+//  Created by Daye on 2021/07/29.
+//
+
+import Foundation
+
+public struct Address{
+    let shopName: String
+    let streetName: String
+    let streetNumber: String
+    let depthOneAddress: String
+    let depthTwoAddress: String
+    let depthThreeAddress: String
+}
